+++ tests/test_ux_features.py
@@ -345,10 +345,10 @@ class TestVersionUpdated:
     """Test that version is correct"""
 
     def test_version_is_3_3_7(self):
-        """Test that version is 3.3.7"""
+        """Test that version is 3.3.8"""
         from cja_auto_sdr.generator import __version__
 
-        assert __version__ == "3.3.7"
+        assert __version__ == "3.3.8"
 
 
 class TestFormatAutoDetection: